Added configuration to enable/disable URL and User Agent checking

+++ src/Http/Middleware/BlockMaliciousUsers.php
@@ -12,25 +12,41 @@ use Closure;
 class BlockMaliciousUsers
 {
 
+    /**
+     * @var \Illuminate\Config\Repository|\Illuminate\Contracts\Foundation\Application|mixed
+     */
+    protected mixed $checkForMaliciousUrls;
+
+    /**
+     * @var \Illuminate\Config\Repository|\Illuminate\Contracts\Foundation\Application|mixed
+     */
+    protected mixed $checkForMaliciousUseragents;
+
+    public function __construct()
+    {
+        $this->checkForMaliciousUrls = (bool) config('laravel-blocker.url_detection_enabled');
+        $this->checkForMaliciousUseragents = (bool) config('laravel-blocker.user_agent_detection_enabled');
+    }
+
     public function handle($request, Closure $next)
     {
         $requestIp = request()->ip();
         $blockedIp = BlockedIp::where('ip', $requestIp)->first();
 
         // Is this a blocked IP that has expired?
-        if ($blockedIp && $blockedIp->hasExpired()) {
+        if ($this->checkUrlsOrAgents() && $blockedIp && $blockedIp->hasExpired()) {
             BlockedIp::where('ip', $requestIp)->delete();
             $blockedIp = null;
         }
 
         // Is this a blocked IP?
-        if ($blockedIp) {
+        if ($this->checkUrlsOrAgents() && $blockedIp) {
             throw new BlockedUserException(__('You have been blocked'), 401);
         }
 
         // Does this URL contain a malicious string?
         // @see config/config.php
-        if (LaravelBlocker::isMailicousRequest()) {
+        if ($this->checkForMaliciousUrls && LaravelBlocker::isMailicousRequest()) {
             // Store blocked IP
             BlockedIp::updateOrCreate(['ip' => $requestIp]);
 
@@ -39,7 +55,7 @@ class BlockMaliciousUsers
 
         // Does the request come from a malicious User Agent?
         // @see config/config.php
-        if (LaravelBlocker::isMaliciousUserAgent()) {
+        if ($this->checkForMaliciousUseragents && LaravelBlocker::isMaliciousUserAgent()) {
             // Store blocked IP
             BlockedIp::updateOrCreate(['ip' => $requestIp]);
 
@@ -47,5 +63,16 @@ class BlockMaliciousUsers
         }
 
         return $next($request);
+    }
+
+    /**
+     * @param mixed $checkForMaliciousUrls
+     * @param mixed $checkForMaliciousUseragents
+     *
+     * @return bool
+     */
+    protected function checkUrlsOrAgents(): bool
+    {
+        return $this->checkForMaliciousUrls || $this->checkForMaliciousUseragents;
     }
 }
